+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## `transcriptic` Changelog
 
 ## Unreleased
+- Updated `Run` class to return all containers used in a run as a data frame, accessed through `Run.containers`
+
 ## v3.3.1
 Fixed
 - Updated `transcriptic runs` route to reflect reality
-- Updated `Run` class to return all containers used in a run as a data frame, accessed through `Run.containers`
 
 ## v3.3.0
 Added