+++ CHANGELOG.md
@@ -1,12 +1,11 @@
 ## `transcriptic` Changelog
 
 ## Unreleased
-
+## v3.4.0
 Added
 - `transcriptic select_org` in CLI now allows you to switch organizations without re-authenticating
 -  `User-agent` information to headers
 -  `Run.containers` to return a list of containers used within the run
-
 
 ## v3.3.1
 Fixed